+++ src/main/java/com/chooz/post/application/PostQueryService.java
@@ -4,11 +4,10 @@ import com.chooz.comment.domain.CommentRepository;
 import com.chooz.common.dto.CursorBasePaginatedResponse;
 import com.chooz.common.exception.BadRequestException;
 import com.chooz.common.exception.ErrorCode;
-import com.chooz.post.application.dto.PostWithVoteCount;
 import com.chooz.post.domain.PollChoice;
 import com.chooz.post.domain.Post;
 import com.chooz.post.domain.PostRepository;
-import com.chooz.post.presentation.UpdatePostResponse;
+import com.chooz.post.presentation.dto.UpdatePostResponse;
 import com.chooz.post.presentation.dto.AuthorDto;
 import com.chooz.post.application.dto.FeedDto;
 import com.chooz.post.presentation.dto.FeedResponse;
@@ -23,7 +22,6 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
-import org.springframework.data.domain.SliceImpl;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
